Add files via upload
Hypothesis: Does depression predict low GPA? Is this relationship mediated by sleep?

Analysis: Multiple Linear Regression

(Using post data)

Do higher levels of depression predict lower GPA?
If it does, how much of this relationship is accounted for by sleep?

Three Separate Analyses:
Does depression predict GPA
Does depression affect sleep
When we control for sleep, what happens to the relationship between depression and GPA

Methods:

Design: The study will be a within-subject design, in which all participants will be exposed to both levels of stress before and after midterm exams. This will allow for the study to be  completed using a smaller number of participants with less chances of error associated with individual differences. Each participant will serve as their own baseline.
Inclusion Criteria: The Dartmouth University students who are enrolled in COSC 065 course over a 10 week period in the spring term of 2013.
Exclusion Criteria:
Participants that have missing data, erroneous data such as 2SD from the mean, or overly consistent responses will be excluded.
Sample
The sample size will comprise Dartmouth University students for whom the Independent, dependent and Mediator variables are available. 

Variables (from the StudentLife Dataset):

Independent variable: Depression among the participants will be the independent variable because it  is expected to vary before and after midterm exams which can affect their academic performance throughout the spring term

Dependent Variable :
Academic Performance of participants will be assessed using the term GPA and cumulative GPA which is expected to be affected by their level of depression.



Mediator Variable:
Sleep Quality and Duration will be included as a mediator variable which is influenced by depression and affects academic performance.


Materials
Depression among participants will be measured using the Patient Health Questionnaire(PHQ-9) which is the major depressive disorder (MDD) module of the full PHQ. It is used to provisionally diagnose depression and grade severity of symptoms in general medical and mental health settings. It has a Cronbach alpha of .89, and excellent test-retest reliability. It scores each of the 9 DSM criteria of MDD as “0” (not at all) to “3” (nearly every day), providing a 0-27 severity score. PHQ-9 was chosen because it objectively determines severity of initial symptoms, and also monitors symptom changes and treatment effects over time. 

The Positive and Negative Affect Schedule (PANAS) will be used to assess the affect of the participants, which comprises two mood scales, one measuring positive affect and the other measuring negative affect. It has 10 descriptors which are used for each Positive Affect (PA) scale and Negative Affect (NA) to define their meanings. The participants in the PANAS are required to respond to a 20-item test using a 5-point scale that ranges from “very slightly” or not at all (1) to extremely (5). The scales have been shown to be highly internally consistent, largely uncorrelated, and stable at appropriate levels over a 2-month time period.

The Pittsburgh Sleep Quality Index (PSQI) will be used to assess the sleep quality and disturbances over a 1-month time interval  among participants. It is a self-rated questionnaire and has a high test-retest reliability and a good validity for patients with primary insomnia.